refactor(avatar): extract AvatarCard component
